Make yarpc module not specific to thrift
